Refactor: Improve data handling, code quality, and documentation
This commit introduces several enhancements to the portfolio website:

- Data Management:
  - Refactored the handling of education data (/api/education) by moving
    hardcoded degree and certification information into separate, dedicated
    data files (`app/shared/data/educationDegrees.ts` and
    `app/shared/data/educationCerts.ts`). The API route now imports
    this data, improving separation of concerns and maintainability.
    This restores the original education data you provided.

- Code Quality & Readability:
  - Removed a development `console.log` statement from the Credentials
    component.
  - Improved accessibility by adding `aria-hidden="true"` to the decorative
    `FaAngleDown` icon in the Intro component.
  - Added TSDoc comments to document props for key reusable components,
    including `ProjectCard`, `IntroCard`, `SchoolCard`, and `CertBadge`,
    enhancing developer understanding.

- Project Structure:
  - Retained the `achievements` field within the project data structure
    for future population, as per your request.

- Documentation:
  - Updated `README.md` to reflect the data refactoring (marking the
    relevant roadmap item as complete) and to add a note about the
    recent focus on code quality and optimization efforts.

- Image Optimization:
  - A review confirmed that `next/image` is already being utilized
    effectively for static image assets, so no changes were needed in this area.
